Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> I just noticed that this does (or may) change the logic depending on the existence of a delegate. Is that on purpose? The commit message doesn't really convey this.

You're right, there are in fact behavior changes that I had missed. This is mentionned in the original commit message for this change:

https://github.com/ahenriksson/haiku/commit/34e2a649b5c64cf263b5bc620407c17b59d76a23

I will restore that commit message for better clarity of what's going on.

Sorry for missing this when I tried to clean up this code and get it ready for review.

Patch-set: 10
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>