Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Are this bug (#15778) related to code in our repository and applying this

It is related to Haiku code so it behaves differently than documented in BeBook, but it is unrelated to this patch. I just give an example of breaking BeOS compatibility that really break important applications some of them are closed source and can't be changed.

I don't know any applications that don't work because of moved message dispatching code and it can't be fixed. Don't calling base class MessageReceived is broken anyway and should be fixed because it breaks scripting, it is documented in BeBook.

Patch-set: 1
Reviewer: Gerrit User 1000172 <1000172@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>