Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Case insensitivity introduce mess and increase risk of hard to investigate troubles (ICU etc.) if there are no good reason for it (case-insensitive FS for example).

You should not be able to have 2 different decorators with the same name except case, e.g. Default and default. This has nothing to do with the file system. This fixes a bug selecting the Default decorator in the decorator pop up in Appearance. It’s not a big deal but I don’t see how this introduces a mess. Decorator names are not currently localized and app server would need some work here to make it so since it always looks for the English word “Default” in order to select the Default decorator. This patch doesn’t change that and I’m not sure how this could be fixed.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>