Update patch set 1
Patch Set 1:

I think the copyright of Haiku is the only place where we want to mention Haiku even in distributions?

The rules are "you can't name your distribution Haiku or something similar, but you must properly credit that it is based on Haiku and reproduce the exact copyright notice"

This isn't an "unspoken" rule at all, it is what the MIT license says (for the copyright crediting part) combined with the trademark usage guidelines (for needing to name your distro differently)

That's what the existing code was doing and I don't think there is a need for more text here? We are perfectly fine with people acknowledging that their work is based on Haiku or pretty close to it. And Haiku inc may even grant exceptions to the trademark rules in some cases.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>