Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think the copyright of Haiku is the only place where we want to mention Haiku even in distributions?

Got it.

> The rules are "you can't name your distribution Haiku or something similar, but you must properly credit that it is based on Haiku and reproduce the exact copyright notice"

Technically, my change pretty much achieves it because credit to Haiku is automatically added (including a slightly paraphrased version of the obligatory notice, per the Guidelines) if the distribution is unofficial.

> This isn't an "unspoken" rule at all, it is what the MIT license says (for the copyright crediting part) combined with the trademark usage guidelines (for needing to name your distro differently)

I issued a correction before I read the comment.

> That's what the existing code was doing and I don't think there is a need for more text here?

I tried to follow what humdinger sort of did in FirstBootPrompt earlier and made a change before I read your comments.

> We are perfectly fine with people acknowledging that their work is based on Haiku or pretty close to it.

> And Haiku inc may even grant exceptions to the trademark rules in some cases.

Thought of that, so I just basically added "Haiku (unofficial)" by default. I can take a step back with the text if that's wanted, but I think that this approach in itself should work fine?

Patch-set: 1