Update patch set 2
Patch Set 2:

I pretty much thought of the same things as you did and I was particularly bothered by the fact that *Distroname* looked pretty ugly in practice and I'd like to heavily consider changing that behavior in FirstBootPrompt as well.

After testing the change in a virtual machine, I also noticed that there were some typos/forgotten spaces etc., so I decided to combine some of my aforementioned ideas and push a new change.

"(based on Haiku)" takes up less space and should be more translatable than "(unofficial Haiku distribution)". (e.g. in German, that'd have to be "(basiert auf Haiku)")

None of this will even appear if the build is official and it just mostly solves the burden that the developer coming up with their own distribution of Haiku for them in a way that we /should be/ happy about.

Patch-set: 2