Update patch set 2
Patch Set 2:

> The distributor is responsible to change that into something appropriate.

Yeah, and it's really just the distribution string and everything else is pretty much ready and looks aesthetically nice without violating that whole Haiku spirit.

> It might be nice to have the name passed through a macro definition

Good idea, not sure if I should do this over here because that'd go beyond the scope of what I'm trying to change here. I'd also have to include FirstBootPrompt and add the macro, which is something that I am still slightly unfamiliar with despite some recent efforts of mine to try and change that.

> with the build system aborting if it is not passed when building nonofficial images

That will be particularly VERY painful for contributors that do not intend to distribute the .iso's either way. My motivating factor was that screenshots looked slightly off with the "Haiku" string and that both "Walter" and "*Distroname*" made it way worse as far as being able to recognize the screenshot. People keysmashing in the ./configure command or entering strings ranging from "Haiku" all the way up to "God is dead" or something edgy like that will prove to be counterproductive.

Patch-set: 2