Update patch set 3
Patch Set 3:

I tried to go through the rabbit hole of trying to figure out how to add a macro based on other definitions in `build/jam/images/definitions/common-tail` and `build/jam/BuildSetup`, but the values of these always end up in shell scripts.

The only sort of solution that I could figure out using Jam itself would be something like /etc/os-release (https://www.freedesktop.org/software/systemd/man/os-release.html) in certain Linux distributions, but I really feel like something like that should probably be some sort of a method exposed to the Haiku APIs.

I think I'm sort of convinced that it would be a good idea to leave this topic alone for the time being, because I fear that this will get excessively convoluted. We really should implement this change and do something similar in FirstBootPrompt before implementing something like this - I do believe that replicating this entire process for every single program that's wondering whether the version of Haiku that's currently being used is "official" is annoyingly tedious (or even what the name of the system is).

Patch-set: 3