Update patch set 3
Patch Set 3:

You are really over complicating this. When building in default mode we had made enough debranding already and the existing thing was fine: the system is referred to as "the system" or "this system" where appropriate.

You are the one introducing the not very convincing *distroname* thing which I think we don't want. "Try out the system" would have worked just fine there as we did elsewhere.

That's all the effort we are willing to pu into this. The message we want to convey is, think really hard before you do your own distro and add your own branding, because that's something we allow (it's opensource), but not encourage (it's not Linux with it's famous Hannah Montana Linux distribution). And so, we make it easy to protect our own trademark, but if you want to add your own branding, you are on your own and you will need to put in some effort. We will not do this work for you.

Patch-set: 3