Update patch set 3
Patch Set 3:

> You are really over complicating this.

That was Max's idea, I was the one that said 'I did look into that idea you recommended, let's not make this too complicated'.

> When building in default mode we had made enough debranding already and the existing thing was fine: the system is referred to as "the system" or "this system" where appropriate.

Not in the existing part that I actually modified.

> You are the one introducing the not very convincing *distroname* thing which I think we don't want. "Try out the system" would have worked just fine there as we did elsewhere.

I think you are wrong here. I did not introduce this.

I just tried to do what was being done in FirstBootPrompt and I'll repeat that I realized that it didn't look good at all, which is why I tried taking a different approach instead.

Unofficial builds of Haiku were referred to as Walter until recently, that changed everywhere (see: FirstBootPrompt, where that *Distroname* thing is implemented already) apart from the distro version.
 
> That's all the effort we are willing to put into this. The message we want to convey is, think really hard before you do your own distro and add your own branding, because that's something we allow (it's opensource).

Alright, so I'll just do what I proposed a bit earlier and take a step back with all the text additions. I'll keep the label that says "(based on Haiku)" if a user supplies a different distribution name, or a label that says "Haiku (unofficial)".

I'm just going to reiterate that if a person, for any reason, were to come up with a custom distribution and did not follow the Guidelines in a manner that's not malicious, that would cause problems on our end, which is a bit weird, considering that we don't even want to provide support for the supposed operating system's users. The goal is to put an additional burden on the developer, but aren't the supposed legal threats and bureaucratic everything an even bigger waste of time that will allocate *our* precious time instead?

Patch-set: 3