Update patch set 4
Patch Set 4:

> I'm just going to reiterate that if a person, for any reason, were to come up with a custom distribution and did not follow the Guidelines in a manner that's not malicious, that would cause problems on our end, which is a bit weird, considering that we don't even want to provide support for the supposed operating system's users. The goal is to put an additional burden on the developer, but aren't the supposed legal threats and bureaucratic everything an even bigger waste of time that will allocate *our* precious time instead?

I still don't understand what part of AboutSystem is not following the guidelines before this change. It is, as I have already said, the only place where Haiku MUST be mentioned and credited. This is the copyright notice for Haiku. If you make a distro based on Haiku, you must keep this and you must keep it unchanged.

If you make changes and put them under the MIT license, you must also add your own copyright notice. But you must not change this one instance of "Haiku" to "Haiku (unofficial)" or "MyDistro (based on Haiku)". You must add a separate section with a separate copyright notice. And this is something we can't automate, especially because what you must do exactly, depends on the license you pick. You could use CC-0 and have nothing to do

Patch-set: 4