Update patch set 4
Patch Set 4:

> I still don't understand what part of AboutSystem is not following the guidelines before this change. It is, as I have already said, the only place where Haiku MUST be mentioned and credited. This is the copyright notice for Haiku. If you make a distro based on Haiku, you must keep this and you must keep it unchanged.
 
> If you make changes and put them under the MIT license, you must also add your own copyright notice. But you must not change this one instance of "Haiku" to "Haiku (unofficial)" or "MyDistro (based on Haiku)". You must add a separate section with a separate copyright notice. And this is something we can't automate, especially because what you must do exactly, depends on the license you pick. You could use CC-0 and have nothing to do

... Apologies, I just had a very hard time understanding where you were coming from because I studied the AboutSystem interface from an entirely different perspective. Thanks for explaining patiently. So wait, just to get it straight, the entire "Haiku" section is a section referring to Haiku, and "Haiku" is not a title. It's not just the name of the operating system, it's just the beginning of a list of the credits *for* Haiku and nothing else.

I thought that the current layout was supposed to be like this:

```
# Name of the operating system

<p>Copyright notice for Haiku + Website</p>
<p>Additional copyright notices</p>

---

## Credits for the other supposed operating system.
## Credits for Haiku
```

I didn't know that the "Haiku" header was actually *a credit* - I thought it was the distribution name, which is why I saw the entire situation as "AboutSystem will make the supposed distro or fork of Haiku non-compliant as soon as the person makes the fork". When you brought up the "copyright notice", I thought that you were solely referring to the "(c) 2008-2021 Haiku project" line, and just that.

I now understand that a modified version of AboutSystem is supposed to look like this:

```
# My header and my own copyright stuff

<p>Whatever</p>

# Haiku header and the rest of the copyright stuff

<p>Copyright notice for Haiku + Website</p>
<p>and so on and so forth</p>
```

I'll try to repurpose some of the changes I proposed here for FirstBootPrompt instead. The entire "Don't modify anything from the entire AboutSystem credits section" should probably be better clarified, but I've already made 15 commits to the website that are outstanding and this seems like a fairly big and convoluted change for me to handle.

Patch-set: 4