Update patch set 1
Patch Set 1:

Why move things to a .h file that is not used anywhere else? Do you have further plans or are you applying a generic "class definitions should go in .h files" rule that we don't usually enforce?

Also, yes, you have used spaces (please configure your editor correctly), you have also added an extra file from VS Code that should not be there (please configure your gitignore settings), and you have submitted a change twice, which should not happen if you use the instructions in our "getting sourcecode" page to set up the git commit hook that will automatically set a change-id to your local commits to avoid this kind of mistake.

I'm sorry to be a bit blunt but I feel this is a lot of noise for little added value. Please take the time to configure your tools properly, review your changes before submitting them, and consider wether the changes are needed in the first place, and if they are, take the time to explain why in the commit message.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>