Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Also, yes, you have used spaces (please configure your editor correctly)

As mentioned privately, this only seems to have applied to one specific line in the Copyright section.

> I'm sorry to be a bit blunt but I feel this is a lot of noise for little added value. Please take the time to configure your tools properly, review your changes before submitting them

I tried to remove the VS Code file but the fact that I kept the IDE open basically put the file back in place. Won't happen again, I hope.

> and consider whether the changes are needed in the first place, and if they are, take the time to explain why in the commit message.

Apart from readability, I had a random idea of taking advantage of the AboutSystem's functions for an experimental screensaver that would act like an animated title sequence a while back and figured that taking advantage of AboutSystem would be a decent strategy so that I don't have to focus on parsing the "Credits.h" file. AboutView::_CreateCreditsView() seemed useful for that, particularly because of the way it already does most of the formatting, as well as AboutView::AddCopyrightEntry().

So I'm also making these changes for the sake of taking advantage of stuff in the system that's already pretty much ready for fun. I'm not sure whether or how I could explain that in the commit message.

Patch-set: 1