Update patch set 3
Patch Set 3:

> So I'm also making these changes for the sake of taking advantage of stuff in the system that's already pretty much ready for fun. I'm not sure whether or how I could explain that in the commit message.

This idea dates back to the Google Code-In 2019 competition, where one of the tasks was to come up with a screensaver from scratch. I eventually dropped the idea because it got way too painful to deal with parsing "Credits.h" properly, which AboutSystem takes care of. I'm not sure if it's a good idea to take advantage of something that isn't a part of the Haiku API, but, then again, I'd still have to take advantage of AboutSystem either way, because it contains "Credits.h" too, and that copying and pasting that file would mean that it would quickly get pretty much outdated.

Patch-set: 3