Update patch set 7
Patch Set 7:

I'm really under a heavy time constraint right now, which is clearly showing because I made a bunch of typos, re-included an .idea folder and messed up the cases. I'll take a break after this change is good to go because this probably not very helpful. Apologies for wasting your (and every other reviewers') time.

Patch-set: 7