Update patch set 3
Patch Set 3:

(1 comment)

My goal is to be able to use excerpts (or the entirety) of the Credits the way they have been processed by AboutWindow for a variety of fun purposes, which I've stated earlier. There's a slim chance that I may choose to make some private functions public later on.

I'm also not sure how I could possibly split this file in smaller parts. It's true that having three MessageReceived() functions in a single source file seems excessive, but it's not like I can put a large sum of the logic in a different source file and follow the "Use a source file for managing the window, and another one for the logic", because most of the code affects how the windows operate. The only exception to that would probably be AboutView::AboutView(), but it still has a lot of window-related logic in it.

Patch-set: 3