Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> My goal is to be able to use excerpts (or the entirety) of the Credits the way they have been processed by AboutWindow for a variety of fun purposes, which I've stated earlier. There's a slim chance that I may choose to make some private functions public later on.

Personally I would work on this from the other end: start writing your screensaver, and see which part(s) of the code you need for it to work. Move these to a separate file, and possibly to a separate class, so you can reuse them there easily.

Patch-set: 3