Update patch set 4
Patch Set 4:

(6 comments)

I think that the only part that should be made public has to be CreateCreditsView(). I also thought of making other internal functions and variables public (such as `_AdjustTextColors()`), but, at this point, I fear that if the developer feels the need to modify the colors and everything, they might as well just write their own AboutView::AboutView() function instead of finding ourselves in a situation where an application would be in a position to cause definition-related conflicts.

Patch-set: 4