Update patch set 1
Patch Set 1:

set_ac() sets the AC flag in eflags/rflags, which is then pushed in the iframe by the CPU whenever a fault happens. After handling the exception, the fault handler is runned with the faulted iframe (with a new frame->ip), which still has the AC flag set. We need to unset the flag after the call to longjmp and the return to the setjmp call location. Otherwise the AC flag is set until the syscall returns to userland.

Patch-set: 1