Update patch set 1
Patch Set 1:

Ping! :) Just to clarify from before, in addition to not being able to get BMessageRunner to work, I believe that using BMessageRunner for the purpose of animation is just not ideal. This is due to the message sending relying on the registrar which creates additional overhead. We could really use a "BTimer" class of sorts.

Patch-set: 1