Update patch set 19
Patch Set 19:

Would be a good idea to continue the review from Patchset 15 (since there was only more confusion after it).

I have tried to fix the multiple leaks that we had. As suggested in early reviews multiple times, BAutoDeleter could be used. But, no documentation on the could be found (couldn't find it anywhere in the source code as well). Hence, I've covered the cases where there could have been a memory leak.

fExtents is a member attribute to the class because it is used in a later patch. It is used when we implement GetNext(), as mentioned in the patch desc.

Patch-set: 19
Reviewer: Gerrit User 1000183 <1000183@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>