Update patch set 19
Patch Set 19:

> Patch Set 19:

> Just to be more clear, I've covered the cases by manually deleting 'extentsWrapped'. Until someone can figure out where we can find BAutoDeleter or any alias function we might need to do manual work like we are doing now.

Autodeleter can be found in:
https://git.haiku-os.org/haiku/tree/headers/private/shared/AutoDeleter.h

Usage of it is relatively simple, and mainly depends on if you're trying to delete an object, array or simple allocation. For e.g. objects, you would instantiate it with something like: ObjectDeleter<YourObjectType> deleter(yourObjectPointer, true);

And in the success case where the function completes and hands ownership of it over, deleter.Detach(); must be called to release it from the deleter. This way, all failure paths where the deleter still has the object attached to it will automatically delete the object on function return, as part of cleaning up the stack.

Patch-set: 19