Update patch set 1
Patch Set 1:

Hi,

I have a problem with the fact that TextRect is now computed only in FrameResized. One of my applications does something like this:

- Create a new BTextView (using the layout based constructor, so the view initial rect is empty)
- Add some text to it
- Get the TextRect
- Use SetExplicitMinSize / SetExplicitMaxSize to set the window size to match the TextRect

In my case I use the text view to have non-editable, formatted text (it is used in a BAlert-like window where I need bold text, etc).

With the current version of the code in the master branch, I get an empty TextRect and find myself setting my view to have a width of 0.

What would be the best way to create a view like this? With a fixed size appropriate to fit the text?

I have not tested yet if your latest changes fix this (not having Gerrit working in Haiku makes my life a bit harder there).

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>