Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I have a problem with the fact that TextRect is now computed only in FrameResized. One of my applications does something like this:

When word-wrap is off FrameResized() changes the size of text rect to match the width of the longest line of text. If word-wrap is on it resizes the text rect width to match the width of the text view instead. The height of the text rect is always set to the height of the text.

However, the text rect is not only set in FrameResized(), there are several other places it is set such as SetTextRect() and SetText().

> - Create a new BTextView (using the layout based constructor, so the view initial rect is empty)
> - Add some text to it

Using SetText()? That method should resize the text rect to fit the text. The only way I could see that method not doing this is if the text view width is 0 at the point that SetText() is called. See why below.

> - Get the TextRect
> - Use SetExplicitMinSize / SetExplicitMaxSize to set the window size to match the TextRect

Using the layout version of the BTextView constructor should set the initial text rect to the text view Bounds(). I'm guessing that because there is no initial text it is setting both the text view and text rect width to be 0.

One way to fix this is to use SetTextRect() to set the text rect to something close to the size you want it to be. Another way is to set the text view width to be something greater than 0 using SetExplicitMinSize().

> In my case I use the text view to have non-editable, formatted text (it is used in a BAlert-like window where I need bold text, etc.).

Being non-editable should not change how the text rect resizes except that it doesn't use the default insets because there is no visible text view boundary in this case.

> With the current version of the code in the master branch, I get an empty TextRect and find myself setting my view to have a width of 0.

What is happening is that the initial text view width (and thus text rect width) is 0 so when _RecalculateLineBreaks() is called to resize the text rect (I assume by SetText()) to fix the text it sees a 0 width and exits. This functionality in _RecalculateLineBreaks() has caused a lot of issues like this.

https://git.haiku-os.org/haiku/tree/src/kits/interface/TextView.cpp#n3802
^ This is the line that exits out of resizing the text view when text rect width is set to 0.

> What would be the best way to create a view like this? With a fixed size appropriate to fit the text?

Use SetTextRect to set the text rect to something reasonable for the text or use SetExplicitMinSize() to set the width (and height) to something greater than 0.

> I have not tested yet if your latest changes fix this (not having Gerrit working in Haiku makes my life a bit harder there).

The FrameResized() code was ported from StyledEdit with some performance improvements and my more recent patch sets don't significantly alter this behavior so I imagine it would behave the same way on my latest patch.

The main problem here is that once the text rect width is set to 0 it will stay 0 until it is explicitly resized (using SetTextRect() or by explicitly resizing the text view). We could alter this behavior but the reason that we made it work this way was to avoid performance problems from resizing the text rect over and over again in AboutSystem credits view.

Patch-set: 1