Update patch set 1
Patch Set 1:

> The main problem here is that once the text rect width is set to 0 it will stay 0 until it is explicitly resized (using SetTextRect() or by explicitly resizing the text view). We could alter this behavior but the reason that we made it work this way was to avoid performance problems from resizing the text rect over and over again in AboutSystem credits view.

The problem in AboutSystem was with ReclaculateLineBreaks trying to fit everything in a text rect with a width of 0. When trying to do this it would attempt to insert a linebreak after every character, and it does this in a quite inefficient way.

The fix that was implemented is "don't compute anything if the textrect is 0". But in my case, I shouldn't need this: I have word wrapping turned off, so I shouldn't get into this slow case. There are no extra linebreaks to insert.

Shouldn't we set the initial text rect to something very wide instead? This would give the best chance of computing the correct linebreaks, instead of doing nothing at all. I can do this in my app, but I think this use case (setting some text, getting a text rect, and only then, setting the view size) is common enough. Ideally I should be able to remove my explicit min/max size and the view will size itself (maybe this already works now, I will try that).

Patch-set: 1