Update patch set 2
Patch Set 2:

This actually goes agains the HIG, which says this in the current version:

"An ellipsis is a series of 3 dots (…) used to tell the user that a control, often a menu item or button, will require further input. For example, a menu item named "New…" will display a window which has the title "New". However, if creating a new document does not require showing a window, then an ellipsis should not be used."

The About menu does not require further input. When the about window opens, the action of the menu is complete. So there is no need for an ellipsis here. This was changed from a previous version of the HIG which stated "an ellipsis is needed if the menu will open a window". This is different from what is done on other operating systems.

See hrev52517 where this change was made to the document.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>