Update patch set 1
Patch Set 1:

Thanks for your patch!

I'm not sure if that's really an issue here, but it would be nicer if the system wouldn't have to try to parse an URL in case it's really a path. IOW I would prefer to move the 'url' variable completely into the 'else' case.

Patch-set: 1
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>