Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I don't think that breaking haiku_loader <-> kernel protocol just before release is a good idea. Many users will get broken UEFI boot and it will give bad impression.

My guess is not as many users are using UEFI + Haiku R1 / Beta 2.

We're going to see more and more people leveraging UEFI in the future... so the longer it takes for us to apply this patch the harder it will be.

Which gets to the core of my argument... if we want this we need to apply it *NOW*.  Otherwise we shouldn't apply it at all and live with the "less than idea" layout as it stands today.

Patch-set: 4
Reviewer: Gerrit User 1000000 <1000000@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>