Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Which gets to the core of my argument... if we want this we need to apply it *NOW*.  Otherwise we shouldn't apply it at all and live with the "less than idea" layout as it stands today.
> > 
> > If breaking protocol, I think that it is better to completely redesign kernel args to use key-value pairs only (like Zircon) instead of hardcoded fields.

We already have at least two key-value based things in the kernel args where it's easy to add more data:
- The driver_settings (which is populated with data from the driver settings files and changes made from the boot menu, but could store more things, it is a linked list of driver_settings_file structures each with a name and some data in "driver settings" format)
- The boot_volume (it is populated by the boot loader in KMessage format and currently contains information about the chosen boot volume)

I think it should be possible to reuse one of these.

The reason some data is passed outside of these is that we need to use it very early in kernel initialization (before we have a heap, etc) and it would be unconvenient to access and parse complex data then. So the idea is that the data required to do the kernel early init is passed statically, and everything else could be passed through these structures.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>