Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > The reason some data is passed outside of these is that we need to use it very early in kernel initialization (before we have a heap, etc) and it would be unconvenient to access and parse complex data then. So the idea is that the data required to do the kernel early init is passed statically, and everything else could be passed through these structures.
> 
> Not sure I understand if that is a problem? Doesn't seem like that should be hard? How does Zircon pull it off?

There is not really a problem, but it makes early kernel code more complicated and that is not an easy place to debug. So the current solution is not that bad: some simple fixed arguments for the early code, and then use higher level things once we have initialized a few things and can run code with debug output, fault handlers set up, etc.

I'm not saying we should not change this, using a key-value system for everything has this limitation, but is simpler for other reasons. And with us supporting more and more diverse hardware and boot platforms, we have new needs that we didn't have when this code was first written.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>