Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Also unsupported in my laptop.
> 
> Anywhere to get VBE/FP doc? I've only found a handful of other projects using it and a reference in vbecore just assigning function 11h and saying there was still no standard.

I checked the VESA website at https://vesa.org/vesa-standards/ (registration needed to access the PDF files) and indeed there is no VBE/FP standard there, and the VBE 3.0 specification says it is a proposal and not yet published.

So I'm not sure where I got the information from.

It seems this part should be removed and we can try to use only the "reduced on" mode from VBE/PM (done in the next commit). This one at least is standard, but a bit less useful (there is only a single "reduced" mode and it is not clear if this just turns the backlight down or does other powersaving things).

For modern displays (possibly including laptops using eDP), another way to set the brightness would be through DDC/CI. This uses the same DDC bus we already use for getting EDID info, and allows to configure the display settings (brightness, contrast, etc). This one is implemented by modern displays, in part because it is required by Microsoft to get a "Windows compatible" approval IIRC. At least my display here supports it 😊

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>