Update patch set 1
Patch Set 1:

Well, I attempted using a BMessageRunner once more, and had no luck. I believe the reason this is happening, is that I'm working from within the registrar itself (which is "command central" for the BMessageRunner system). Not to mention, upon investigating the code of the registrar further, I see that the "shutdown processing system" makes use of the same event queue that handles BMessageRunner events, and likely blocks out other non-shutdown related events. This also disqualifies the usage of Pulse() and SetPulseRate(), as the pulse system makes use of a BMessageRunner in the background (found in BWindow's code).

Once again, while I know that using threads and synchronization is not ideal, I have heavily tested my code to make sure that it is functional. Nonetheless, if there's another timing system that I'm forgetting beyond BMessageRunner and the pulse system, I'd be able to investigate that too. :)

Patch-set: 1