sys/resource.h: add rusage compatibility fields
POSIX defines this structure but specifies only two fields (which we
already implement). However, both the *BSD and Linux have agreed on
some more fields, which are often assumed to be there by applications.

The added size of the rusage struct isn't important for the involved
system calls, the two first fields being referenced by offsets. The
benefice of having compatibility fields is greater as having to patch
every other software at HaikuPorts.