Update patch set 1
Patch Set 1:

So are they just left uninitialized? Or are they guaranteed to be set to 0?

I don't like this approach of just lying to applications about what we support here. It makes it impossible for them to handle the problem properly, no configure check will be able to detect that the fields are there, but in fact not working.

Several of these fields are documented as "unmaintained and always set to 0" in Linux,I guess for these we should not have too much problems.

The remaining ones would be:

ru_maxrss: should be easy to derive from get_area_info? And seems reasonably useful.
ru_minflt/ru_majflt: do we track this per-process already? (number of page faults)
ru_inblock/ru_outblock: do we track this at all? (number of disk IOs)
ru_nvcsw/ru_nivcsw: do we track this at all? (number of context switches)

I think ru_maxrss is the most likely to be used by apps? Can we implement at least that?

For reference, list of affected things in haikuports:
https://github.com/haikuports/haikuports/search?p=2&q=rusage

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>