Update patch set 10
Patch Set 10:

I verified that removing the 0 width check in BTextView::_RecalculateLineBreaks() causes the performance issues described in #5582 to return.

> The problem in AboutSystem was with ReclaculateLineBreaks trying to fit everything in a text rect with a width of 0. When trying to do this it would attempt to insert a linebreak after every character, and it does this in a quite inefficient way.

Layout text view’s set their text rect dimensions to 0 in the constructor because that is the initial Bounds() until the view is added to the layout. Calling SetTextRect() or SetExplicitMinSize() in AboutSystem does not have any effect on the performance problem described in #5582.

I’d like to figure out a way to eliminate the confusion caused by layout text views having 0 dimension bounds initially. The text rect dimensions get fixed up in AttachedToWindow() once the view is attached but if you call TextRect() to get the size of the rect before then it will be invalid and any calculations you base on this will be so as well.

Patch-set: 10
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>